fixed sectors

+++ src/Controller/EnterpriseTypesController.php
@@ -46,6 +46,12 @@ class EnterpriseTypesController extends AppController
      */
     public function add()
     {
+        $session=$this ->request->getSession();        
+        $loggedin_user_role=$session->read('auth-role');
+        if ($loggedin_user_role !== 'admin' &&  $loggedin_user_role!=='super'){            
+            $this->Flash->info(__('Admin privileges required'));
+            return $this->redirect(['action' => 'index']);
+        }
         $enterpriseType = $this->EnterpriseTypes->newEmptyEntity();
         if ($this->request->is('post')) {
             $enterpriseType = $this->EnterpriseTypes->patchEntity($enterpriseType, $this->request->getData());
@@ -68,6 +74,12 @@ class EnterpriseTypesController extends AppController
      */
     public function edit($id = null)
     {
+        $session=$this ->request->getSession();        
+        $loggedin_user_role=$session->read('auth-role');
+        if ($loggedin_user_role !== 'admin' &&  $loggedin_user_role!=='super'){            
+            $this->Flash->info(__('Admin privileges required'));
+            return $this->redirect(['action' => 'index']);
+        }
         $enterpriseType = $this->EnterpriseTypes->get($id, [
             'contain' => [],
         ]);
@@ -92,6 +104,12 @@ class EnterpriseTypesController extends AppController
      */
     public function delete($id = null)
     {
+        $session=$this ->request->getSession();        
+        $loggedin_user_role=$session->read('auth-role');
+        if ($loggedin_user_role!=='super'){            
+            $this->Flash->error(__('Super Admin privileges required'));
+            return $this->redirect(['action' => 'index']);
+        }
         $this->request->allowMethod(['post', 'delete']);
         $enterpriseType = $this->EnterpriseTypes->get($id);
         if ($this->EnterpriseTypes->delete($enterpriseType)) {
